+++ packages/blitz/src/cli/utils/helpers.ts
@@ -1,0 +1,9 @@
+import {readdirSync} from "fs-extra"
+
+export function getPkgManager() {
+  return readdirSync(process.cwd()).includes("pnpm-lock.yaml")
+    ? "pnpm"
+    : readdirSync(process.cwd()).includes("yarn-lock.yaml")
+    ? "yarn"
+    : "npm"
+}