+++ packages/blitz-rpc/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @blitzjs/rpc
 
+## 2.0.0-alpha.56
+
+### Patch Changes
+
+- Updated dependencies [3f20a474]
+- Updated dependencies [abb1ad5d]
+- Updated dependencies [abe2afcc]
+- Updated dependencies [0ac6e171]
+- Updated dependencies [8bcb471a]
+  - @blitzjs/auth@2.0.0-alpha.56
+  - blitz@2.0.0-alpha.56
+
+## 2.0.0-alpha.55
+
+### Patch Changes
+
+- df7cee84: Fix pipe resolver return type
+- Updated dependencies [8f166a5d]
+- Updated dependencies [54a66a95]
+- Updated dependencies [1c809094]
+  - blitz@2.0.0-alpha.55
+  - @blitzjs/auth@2.0.0-alpha.55
+
 ## 2.0.0-alpha.54
 
 ### Patch Changes